Redesign the connection to be just cache. I'll create a 'driver' that will execute the salesforce api request. Add a limit on fetch variables.
